+++ elasticsearch_dsl/__init__.py
@@ -9,6 +9,6 @@ from .mapping import Mapping
 from .index import Index
 from .analysis import analyzer, token_filter, char_filter, tokenizer
 
-VERSION = (0, 0, 4, 'dev')
+VERSION = (0, 0, 4)
 __version__ = VERSION
 __versionstr__ = '.'.join(map(str, VERSION))