+++ Changelog.rst
@@ -6,6 +6,12 @@ Changelog
 0.0.4 (dev)
 -----------
 
+Bug fixes:
+
+ * only document related metadata
+   (``elasticsearch_dsl.document.DOC_META_FIELDS``) can be stored by setting
+   attribute on a document (#58).
+
 0.0.3 (2015-01-23)
 ------------------