+++ Changelog.rst
@@ -6,9 +6,10 @@ Changelog
 0.0.4 (dev)
 -----------
 
- * only document id can now be set directly on a document. Rest of the metadata
-   fields (such as index, version etc) must be stored (and retrieved) using the
-   ``_meta`` attribute (#58)
+ * Metadata fields (such as id, parent, index, version etc) must be stored (and
+   retrieved) using the ``meta`` attribute (#58) on both ``Result`` and
+   ``DocType`` objects or using their underscored variants (``_id``,
+   ``_parent`` etc)
  * query on Search can now be directly assigned
  * ``suggest`` method added to ``Search``
  * ``Search.doc_type`` now accepts ``DocType`` subclasses directly